main.py
import os
from dotenv import load_dotenv

# Imports for LangChain / LangGraph components and local modules
from langchain_core.documents import Document
from src.core_config import load_hf_pipeline, load_hf_embeddings, ResearchRAGState
from src.ingestion import load_documents, get_text_splitter, create_vectorstore, get_retriever
from src.workflow_builder import build_research_rag_graph

# Application entry point for the Agentic Research RAG system.
# This file contains the high-level orchestration signatures and docstring-level comments
# explaining how the runtime should be wired. No implementation logic is included here.


def run_ingestion_pipeline() -> object:
    """
    Ingestion pipeline entry-point.

    Responsibilities (comment-only):
    - Load environment configuration (paths, model IDs, tokenizer name).
    - Initialize a specialized document loader designed for research PDFs (Dedoc or Unstructured).
    - Read the raw research paper found at PDF_INPUT_PATH and convert into a list of
      LangChain `Document` objects while preserving structure (sections, tables, metadata).
    - Instantiate a token-aware `TokenTextSplitter` using the Hugging Face tokenizer
      for the configured LLM model; ensure the splitter uses `from_huggingface_tokenizer`
      so chunk sizes are strictly enforced by token count.
    - Apply the splitter to documents to produce context-preserving chunks.
    - Load the open-source Hugging Face embedding model (EMBEDDING_MODEL_ID) and
      transform chunks into embeddings.
    - Create or persist a ChromaDB vectorstore from the embeddings and return
      a LangChain `Retriever` object configured with the tuned `k` parameter.

    Returns:
        retriever: A configured LangChain Retriever backed by ChromaDB
    """
    pass


def run_rag_chat_loop(compiled_graph: object) -> None:
    """
    Interactive chat loop for running LangGraph compiled workflows.

    Responsibilities (comment-only):
    - Accept user input queries in a loop (CLI or lightweight web UI wrapper).
    - For each incoming query, seed the `ResearchRAGState` with the `user_query`
      and any optionally persisted `chat_history`.
    - Execute the compiled LangGraph workflow and block until a terminal state is
      returned. The graph execution should be fully traced (LangSmith) if
      environment tracing variables are enabled.
    - Print or return the `state['final_answer']` to the caller and optionally
      append the exchange to `chat_history` for multi-turn contexts.

    Args:
        compiled_graph: The compiled/ready-to-run LangGraph instance returned by
                        `build_research_rag_graph`.
    """
    pass


def main() -> None:
    """
    Top-level runtime bootstrap for the Research RAG system.

    Responsibilities (comment-only):
    - Load .env variables using `python-dotenv`.
    - Configure LangSmith tracing environment variables (ensure LANGSMITH_TRACING is 'true').
    - Execute `run_ingestion_pipeline` to create and persist the Chroma index and obtain a retriever.
    - Load the main generation LLM via `load_hf_pipeline` (local Hugging Face pipeline).
      * Note: When targeting TensorFlow, ensure the chosen model has TF-compatible
        weights (e.g., TFAuto* checkpoints). Configure TensorFlow GPU support
        (CUDA/cuDNN) as appropriate and consider `tf.keras.mixed_precision`
        to improve throughput on supported hardware.
    - Call `build_research_rag_graph(llm, retriever)` to obtain a compiled graph.
    - Start `run_rag_chat_loop(compiled_graph)` to accept queries interactively.

    Safety & Operational Notes (comment-only):
    - Users must ensure the chosen LLM and embedding models are available locally
      or can be downloaded with the provided HUGGINGFACEHUB_API_TOKEN.
    - When running locally on GPU with TensorFlow, verify TensorFlow, CUDA and
      cuDNN versions are compatible with your GPU and the installed TensorFlow wheel.
    """
    pass


if __name__ == "__main__":
    main()
