src/ingestion.py
# Ingestion module: document loaders, token-aware splitters, and vectorstore wiring.

from typing import List, Any
from langchain_community.document_loaders import DedocFileLoader, UnstructuredPDFLoader
from langchain_text_splitters import TokenTextSplitter, RecursiveCharacterTextSplitter
from langchain_community.vectorstores import Chroma
from langchain_core.documents import Document
from langchain_core.embeddings import Embeddings


def load_documents(pdf_path: str) -> List[Document]:
    """
    Load and parse PDF research documents with structure preservation.

    Intent and implementation guidance (comment-only):
    - Choose either DedocFileLoader (preferred for table-preserving extraction)
      or UnstructuredPDFLoader depending on system dependencies availability.
    - Ensure table extraction is enabled (e.g., Dedoc with_tables=True) so that
      tables are converted into structured text or JSON blocks that downstream
      chunking can handle.
    - Return a list of LangChain Document objects including metadata about page,
      section headings, and any extracted table content.
    """
    pass


def get_text_splitter(tokenizer_name: str, chunk_size: int = 1024, chunk_overlap: int = 128) -> TokenTextSplitter:
    """
    Instantiate a TokenTextSplitter tied to the Hugging Face tokenizer.

    Implementation guidance (comment-only):
    - Use TokenTextSplitter.from_huggingface_tokenizer(tokenizer, chunk_size, chunk_overlap)
      where `tokenizer` is loaded via `AutoTokenizer.from_pretrained(tokenizer_name)`.
    - Using the HF tokenizer ensures chunk boundaries respect actual token counts for
      the chosen LLM and prevents context-window overflow during generation.
    - If the HF tokenizer is unavailable, fallback to RecursiveCharacterTextSplitter
      with conservative chunk sizes.
    """
    pass


def create_vectorstore(docs: List[Document], embeddings: Embeddings, persist_directory: str = None) -> Chroma:
    """
    Index split documents into a ChromaDB vector store.

    Implementation guidance (comment-only):
    - Convert Document objects into text chunks and embed them using the provided
      embeddings model in batched mode for efficiency.
    - Create a persistent Chroma instance with `persist_directory` set to CHROMA_PERSIST_DIR
      from environment variables to avoid re-ingestion between runs.
    - Store necessary metadata (document id, source, page numbers) alongside vectors
      to enable provenance reporting in generation_node.
    - Return the instantiated Chroma vectorstore object.
    """
    pass


def get_retriever(vectorstore: Chroma, k: int = 4) -> Any:
    """
    Convert a Chroma vectorstore into a LangChain Retriever configured for top-k.

    Implementation guidance (comment-only):
    - Expose parameters like `k`, `score_threshold`, and optionally use hybrid
      search configurations if available.
    - Return a Retriever or a function that accepts a query and returns top-k
      LangChain Document objects used by the retrieval_node.
    """
    pass
